+++ r/R/RcppExports.R
@@ -729,6 +729,14 @@ apply_matrix_double_cpp <- function(mat_sexp, f, row_major) {
     .Call(`_BPCells_apply_matrix_double_cpp`, mat_sexp, f, row_major)
 }
 
+matrix_compute_max_per_row <- function(matrix) {
+    .Call(`_BPCells_matrix_compute_max_per_row`, matrix)
+}
+
+matrix_compute_max_per_col <- function(matrix) {
+    .Call(`_BPCells_matrix_compute_max_per_col`, matrix)
+}
+
 matrix_identical_uint32_t_cpp <- function(mat1, mat2) {
     .Call(`_BPCells_matrix_identical_uint32_t_cpp`, mat1, mat2)
 }